Attach console for container builds

+++ src/shell/stage_launch_cleanup.sh
@@ -89,32 +89,28 @@ zfs_lookup_origin()
 
 cleanup()
 {
-    case "$type" in
-    build)
-        #
-        # Clean up build contexts and bootstrap scripts
-        #
-        rm -fr "${data_root}/instances/${instance}.tar.gz"
-        rm -fr "${data_root}/instances/${instance}.*.sh"
-        rm -fr "${data_root}/instances/${instance}/images"
-        stage_list=$(echo "${data_root}"/instances/"${instance}"/[0-9]*)
-        for d in $stage_list; do
-            umount -f "${d}/root/dev"
+    if [ -f "${data_root}/instances/${instance}.tar.gz" ]; then
+        echo DEBUG: doing build clean up
+            rm -fr "${data_root}/instances/${instance}.tar.gz"
+            rm -fr "${data_root}/instances/${instance}.*.sh"
+            rm -fr "${data_root}/instances/${instance}/images"
+            stage_list=$(echo "${data_root}"/instances/"${instance}"/[0-9]*)
+            for d in $stage_list; do
+                umount -f "${d}/root/dev"
+                case $CBLOCK_FS in
+                ufs)
+                    umount -f "${d}/root"
+                    ;;
+                esac
+            done
             case $CBLOCK_FS in
-            ufs)
-                umount -f "${d}/root"
+            zfs)
+                build_root_vol=$(path_to_vol "${data_root}/instances/${instance}")
+                zfs destroy -f -r "${build_root_vol}"
                 ;;
             esac
-        done
-        case $CBLOCK_FS in
-        zfs)
-            build_root_vol=$(path_to_vol "${data_root}/instances/${instance}")
-            # Recursively remove the ZFS datasets (snapshots and file systems)
-            zfs destroy -r "${build_root_vol}"
-            ;;
-        esac
-        ;;
-    regular)
+    else
+        echo DEBUG: doing regular cleanup
         case $CBLOCK_FS in
         fuse-unionfs)
             umount_reverse_order
@@ -132,8 +128,7 @@ cleanup()
             umount_reverse_order
             ;;
         esac
-        ;;
-    esac
+    fi
 }
 
 kill_jail
